Struct actix_server::ServerBuilder

pub struct ServerBuilder { /* fields omitted */ }

Server builder

Methods

impl ServerBuilder

Important traits for ServerBuilder
impl Future for ServerBuilder type Output = ();

pub fn new() -> ServerBuilder

Create new Server builder instance

pub fn workers(self, num: usize) -> Self

Set number of workers to start.

By default server uses number of available logical cpu as workers count.

pub fn backlog(self, num: i32) -> Self

Set the maximum number of pending connections.

This refers to the number of clients that can be waiting to be served. Exceeding this number results in the client getting an error when attempting to connect. It should only affect servers under significant load.

Generally set in the 64-2048 range. Default value is 2048.

This method should be called before bind() method call.

pub fn maxconn(self, num: usize) -> Self

Sets the maximum per-worker number of concurrent connections.

All socket listeners will stop accepting connections when this limit is reached for each worker.

By default max connections is set to a 25k per worker.

pub fn system_exit(self) -> Self

Stop actix system.

pub fn disable_signals(self) -> Self

Disable signal handling

pub fn shutdown_timeout(self, sec: u64) -> Self

Timeout for graceful workers shutdown in seconds.

After receiving a stop signal, workers have this much time to finish serving requests. Workers still alive after the timeout are force dropped.

By default shutdown timeout sets to 30 seconds.

pub fn configure<F>(self, f: F) -> Result<ServerBuilder> where
F: Fn(&mut ServiceConfig) -> Result<()>,

Execute external configuration as part of the server building process.

This function is useful for moving parts of configuration to a different module or even library.

pub fn bind<F, U, N: AsRef<str>>(
self,
name: N,
addr: U,
factory: F
) -> Result<Self> where
F: ServiceFactory<TcpStream>,
U: ToSocketAddrs,

Add new service to the server.

pub fn bind_uds<F, U, N>(self, name: N, addr: U, factory: F) -> Result<Self> where
F: ServiceFactory<UnixStream>,
N: AsRef<str>,
U: AsRef<Path>,

Add new unix domain service to the server.

pub fn listen_uds<F, N: AsRef<str>>(
self,
name: N,
lst: UnixListener,
factory: F
) -> Result<Self> where
F: ServiceFactory<UnixStream>,

Add new unix domain service to the server. Useful when running as a systemd service and a socket FD can be acquired using the systemd crate.

pub fn listen<F, N: AsRef<str>>(
self,
name: N,
lst: TcpListener,
factory: F
) -> Result<Self> where
F: ServiceFactory<TcpStream>,

Add new service to the server.

Important traits for Server
impl Future for Server type Output = Result<()>;

pub fn run(self) -> Server

Starts processing incoming connections and return server controller.